requirements.txt
Django==4.1
djangorestframework==3.15.1
PyJWT==2.8.0
PyJWT==2.8.0
